feat: add delete confrim board modal

+++ src/components/UI/Board/Board.tsx
@@ -1,8 +1,10 @@
-import React, { FC, MouseEvent } from 'react';
+import React, { FC, MouseEvent, useState } from 'react';
 import cl from './Board.module.scss';
 import { useNavigate } from 'react-router-dom';
 import { useDeleteBoardMutation } from '../../../API/boardsCalls';
 import { useTranslate } from '../../../hooks/useTranslate';
+import { Modal } from '../../../components/Modal/modal';
+import ConfirmModal from '../../../components/Modal/modals/confirmModal';
 
 interface IBoardProps {
   board: {
@@ -18,15 +20,32 @@ const Board: FC<IBoardProps> = ({ board }) => {
   const navigate = useNavigate();
   const [deleteBoard, {}] = useDeleteBoardMutation();
   const T = useTranslate();
+  //Modal manipulations
+  ////////////////////////
+  const [isModalOpen, setModalOpen] = useState(false);
 
+  ///////////////////////////////////////////////////
   const boardOnClick = () => {
     navigate(`/main/${id}`);
   };
 
   const deleteOnClick = (e: MouseEvent) => {
     e.stopPropagation();
-    deleteBoard(id);
+    setModalOpen(true);
   };
+
+  function confirmDeleteBoard(e: React.MouseEvent<HTMLElement>) {
+    e.stopPropagation();
+    const target = (e.target as HTMLElement).closest('input');
+    const value = target?.value;
+    if (value == 'No' || value == 'Нет') {
+      setModalOpen(false);
+    }
+    if (value == 'Yes' || value == 'Да') {
+      setModalOpen(false);
+      deleteBoard(id);
+    }
+  }
 
   return (
     <div className={cl.container} onClick={boardOnClick}>
@@ -37,6 +56,11 @@ const Board: FC<IBoardProps> = ({ board }) => {
       <button className={cl.delete} onClick={(e) => deleteOnClick(e)}>
         {T('Board.delete')}
       </button>
+      {isModalOpen && (
+        <Modal>
+          <ConfirmModal handler={confirmDeleteBoard}></ConfirmModal>
+        </Modal>
+      )}
     </div>
   );
 };
